telegram-bot/requirements.txt
aiogram==3.22.0
aiohttp==3.12.15
pydantic==2.11.7
python-dotenv==1.1.1